Reorganizing files - provide alerts.scss that will work regardless of whether using bower or npm.  Add test coverage.
