+++ .npmignore
@@ -5,6 +5,7 @@
 .npmignore
 .travis.yml
 bower_components
+coverage
 node_modules
 test
 preprocessor.js
