+++ .travis.yml
@@ -1,6 +1,8 @@
 language: node_js
 node_js:
-- '0.10'
+- 4.1
+after_success:
+- npm run report-coverage
 deploy:
   provider: npm
   email: d2ltravisdeploy@d2l.com
