+++ alerts.scss
@@ -1,5 +1,3 @@
 @import 'bower_components/vui-alerts/error';
 
-.app-error {
-	@include vui-alerts-error;
-}
+@import 'src/error';
