+++ package.json
@@ -1,19 +1,21 @@
 {
   "name": "react-vui-alerts",
-  "version": "0.4.0",
+  "version": "0.5.0",
   "description": "React component for displaying alerts",
   "main": "index.js",
   "scripts": {
     "prebuild": "npm run clean -s",
-    "build:css": "node-sass --output-style expanded error.scss dist/error.css && cp ./*.scss ./dist/",
+    "build:css": "node-sass --output-style expanded alerts.scss dist/alerts.css",
     "build:js": "browserify -d -x react -s react-vui-alerts ./index.js | exorcist ./dist/react-vui-alerts.js.map > ./dist/react-vui-alerts.js && npm run build:js:min",
     "build:js:min": "uglifyjs ./dist/react-vui-alerts.js -o ./dist/react-vui-alerts.min.js --source-map ./dist/react-vui-alerts.min.js.map --in-source-map ./dist/react-vui-alerts.js.map",
     "build": "npm run build:css && npm run build:js",
     "clean": "rimraf dist && mkdir dist && rimraf test/coverage",
     "postinstall": "bower install",
+    "report-coverage": "coveralls < ./coverage/lcov.info",
     "test:lint:css": "csslint ./dist/*.css",
-    "test:lint:js": "eslint --ext .js --ext .jsx ./*.js",
-    "test": "npm run build && npm run test:lint:css && npm run test:lint:js && jest"
+    "test:lint:js": "eslint --ext .js --ext .jsx ./src",
+    "test:unit": "jest --coverage",
+    "test": "npm run build && npm run test:lint:css && npm run test:lint:js && npm run test:unit"
   },
   "repository": {
     "type": "git",
@@ -35,16 +37,17 @@
     "react": "^0.13.3"
   },
   "devDependencies": {
+    "babel-jest": "^5.3.0",
     "babelify": "^6.1.3",
-    "babel-eslint": "^4.0.10",
     "bower": "^1.5.3",
     "browserify": "^11.0.1",
     "csslint": "^0.10.0",
+    "coveralls": "^2.11.4",
     "eslint": "^1.2.0",
     "eslint-config-brightspace": "^0.0.2",
     "exorcist": "^0.4.0",
-    "jest-cli": "^0.4.15",
-    "node-sass": "^3.1.2",
+    "jest-cli": "^0.5.8",
+    "node-sass": "^3.3.3",
     "react": "^0.13.3",
     "react-tools": "^0.13.3",
     "rimraf": "^2.4.2",
@@ -61,10 +64,14 @@
     ]
   },
   "jest": {
-    "scriptPreprocessor": "<rootDir>/preprocessor.js",
+    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
     "testDirectoryName": "test",
     "unmockedModulePathPatterns": [
       "react"
-    ]
+    ],
+    "collectCoverageOnlyFrom": {
+      "index.js": true,
+      "src/error.js": true
+    }
   }
 }
