+++ src/error.scss
@@ -1,0 +1,3 @@
+.app-error {
+	@include vui-alerts-error;
+}
